+++ src/mapnik_image.hpp
@@ -56,7 +56,6 @@ public:
     static Local<Value> _premultiplySync(_NAN_METHOD_ARGS);
     static NAN_METHOD(premultiplySync);
     static NAN_METHOD(premultiply);
-    static NAN_METHOD(premultiplied);
     static void EIO_Premultiply(uv_work_t* req);
     static Local<Value> _demultiplySync(_NAN_METHOD_ARGS);
     static NAN_METHOD(demultiplySync);
@@ -91,6 +90,8 @@ public:
     static NAN_SETTER(set_scaling);
     static NAN_GETTER(get_offset);
     static NAN_SETTER(set_offset);
+    static NAN_GETTER(get_premultiplied);
+    static NAN_SETTER(set_premultiplied);
 
     void _ref() { Ref(); }
     void _unref() { Unref(); }